Continúa la implementación del árbol B*. Se modificaron los constructores de los nodos para que reciban el tamaño de los mismos y para que calculen el tamaño mínimo dependiendo de si se trata de un nodo hoja o no. Se terminaron de implentar los métodos buscarpadre, buscarLugar y buscarMenorMayores de la clase BStarTree.

+++ trunk/CapaIndices/Common/BloqueIndice.h
@@ -34,7 +34,7 @@ class BloqueIndice
 		///////////////////////////////////////////////////////////////////////////
 		// Atributo
 		///////////////////////////////////////////////////////////////////////////
-		unsigned int espacioLibre;
+		unsigned short espacioLibre;
 		
 	public:
 		BloqueIndice() {}
@@ -44,12 +44,12 @@ class BloqueIndice
 		///////////////////////////////////////////////////////////////////////////
 		// Getter/Setter
 		///////////////////////////////////////////////////////////////////////////		
-		unsigned int getEspacioLibre() const
+		unsigned short getEspacioLibre() const
 		{
 			return espacioLibre;
 		}
 		
-		void setEspacioLibre(unsigned int espacio)
+		void setEspacioLibre(unsigned short espacio)
 		{
 			this->espacioLibre = espacio;
 		}
